Update to list apis and limit to single controller
