--- looking-at-api-explorer/ApiExplorerExample/ApiExplorerExample/ApiExplorerVisibilityEnabledConvention.cs
@@ -1,25 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.Mvc.ApplicationModels;
-
-namespace ApiExplorerWebSite
-{
-    // Enables ApiExplorer for controllers that haven't explicitly configured it.
-    // This is part of the test that validates that ApiExplorer can be configured via
-    // convention
-    public class ApiExplorerVisibilityEnabledConvention : IApplicationModelConvention
-    {
-        public void Apply(ApplicationModel application)
-        {
-            foreach (var controller in application.Controllers)
-            {
-                if (controller.ApiExplorer.IsVisible == null)
-                {
-                    controller.ApiExplorer.IsVisible = true;
-                    controller.ApiExplorer.GroupName = controller.ControllerName;
-                }
-            }
-        }
-    }
-}
