--- looking-at-api-explorer/ApiExplorerExample/ApiExplorerExample/Controllers/ApiExplorerVisibilityEnabledByConventionController.cs
@@ -1,16 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.Mvc;
-
-namespace ApiExplorerWebSite
-{
-    [Route("ApiExplorerVisbilityEnabledByConvention")]
-    public class ApiExplorerVisbilityEnabledByConventionController : Controller
-    {
-        [HttpGet]
-        public void Get()
-        {
-        }
-    }
-}
