--- looking-at-api-explorer/ApiExplorerExample/ApiExplorerExample/Views/Home/About.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-    ViewData["Title"] = "About";
-}
-<h2>@ViewData["Title"].</h2>
-<h3>@ViewData["Message"]</h3>
-
-<p>Use this area to provide additional information.</p>
