+++ looking-at-api-explorer/ApiExplorerExample/ApiExplorerExample/Views/_ViewImports.cshtml
@@ -1,2 +1,2 @@
-﻿@using WebApplication1
+﻿@using ApiExplorerExample
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
